pyproject.toml
[project]
name = "a2a-holiday-planner"
version = "0.1.0"
# lock to your real runtime so uv doesn't try to satisfy other Python versions
requires-python = ">=3.11,<3.12"


dependencies = [
  "uvicorn",
    "google-adk>=1.2.1",
    "google-generativeai",
  "httpx",
  "python-dotenv",
  "starlette",
  "sse-starlette",
  "langgraph",
  "langchain-core",
  "langchain-openai",
  "a2a-sdk>=0.2.5",
  "crewai[tools]>=0.126.0",
  "pydantic",
  "streamlit>=1.50.0",
]
